Add PREFIX state to support other card readers.
Some card readers prefix the scanned data with a string like
"\!STCARD A " before the % character. This can be specified
in the configuration with the prefixCharacter property. An extra
state in the state machine will eat up characters between the prefix
character and the % character.  The contents of the prefix are
discarded, and are not available to the client.